requirements.txt
flask
flask_sqlalchemy
pymysql
gunicorn
psycopg2
python-dotenv